Tenth Edit

+++ main.py
@@ -8,7 +8,7 @@ def home():
 @app.route('/add/<num1>/<num2>')
 def add_nums(num1, num2):
     if num1.isdigit()==num2.isdigit()==True:
-        res=str(float(num1)+float(num2))
+        res=str(int(num1)+int(num2))
         return render_template('add.html',num1=num1,num2=num2,res=res)
     else:
         return render_template('nonDigit.html')
@@ -16,7 +16,7 @@ def add_nums(num1, num2):
 @app.route('/sub/<num1>/<num2>')
 def sub_nums(num1, num2):
     if num1.isdigit()==num2.isdigit()==True:
-        res=str(float(num1)-float(num2))
+        res=str(int(num1)-int(num2))
         return render_template('subtract.html',num1=num1,num2=num2,res=res)
     else:
         return render_template('nonDigit.html')
@@ -24,7 +24,7 @@ def sub_nums(num1, num2):
 @app.route('/mult/<num1>/<num2>')
 def mult_nums(num1, num2):
     if num1.isdigit()==num2.isdigit()==True:
-        res=str(float(num1)*float(num2))
+        res=str(int(num1)*int(num2))
         return render_template('multiply.html',num1=num1,num2=num2,res=res)
     else:
         return render_template('nonDigit.html')
@@ -35,7 +35,7 @@ def divd_nums(num1, num2):
         if int(num2) == 0:
             res=num2
         else:
-            res=str(float(num1)/float(num2))
+            res=str(float(num1)/int(num2))
         return render_template('divide.html',num1=num1,num2=num2,res=res)
     else:
         return render_template('nonDigit.html')
